src/Lemmatizer-LTN/etl/requirements.txt
google-api-python-client
google-auth
psycopg[binary]>=3
